[mod] CommunityAPI Service 분리, APIError ErrorModel 이동
