+++ pp/pp/Model/ErrorModel.swift
@@ -6,3 +6,14 @@
 //
 
 import Foundation
+
+
+//MARK: - APIError Type
+
+struct APIError: Decodable, Error {
+    let type: String
+    let title: String
+    let status: Int
+    let detail: String
+    let instance: String
+}
