+++ pp/pp/Model/PostModel.swift
@@ -122,14 +122,6 @@ struct CommentRequest: Codable {
    
 }
 
-//MARK: - APIError Type
 
-struct APIError: Decodable, Error {
-    let type: String
-    let title: String
-    let status: Int
-    let detail: String
-    let instance: String
-}
